Fix error in CourseAdminForm.clean()
Addresses Sentry error:
https://sentry.ccnmtl.columbia.edu/sentry-internal/mediathread/group/922/

It looks like this boolean was flipped... maybe there was some confusion
because of the double negative:

	if not 'course_string' not in self._errors:

I think what this piece of code is intending to do is initialize
`self._errors['course_string']` if it doesn't exist.

I added a test to go through this code path.